pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
authors = [
    {name = "楚天尋簫工作室"},
    {name = "ictye", email = "ictyeoffice@gmail.com"},
]
name = "ictye_live_dm"
version = "1.0"
license = {file = "LICENSE"}
readme = {file = "README.md", content-type = "text/markdown"}
description = "a modular danmuk platform"
keywords = ["danmuk","live"]
dependencies = [
    "aiohttp>=3.9.2",
    "aiowebsocket>=1.0.0.dev2",
    "Brotli>=1.1.0",
    "pycryptodome>=3.19.1",
    "PyYAML>=6.0.1",
    "Requests>=2.31.0",
]

[project.urls]
Repository = "https://github.com/northgreen/ictye-live-dm.git"
Issues = "https://github.com/northgreen/ictye-live-dm/issues"
Homepage = "https://github.com/northgreen/ictye-live-dm"

[tool.hatch.build.targets.wheel]
packages = ["src/ictye_live_dm","src/ictye-live-dm/depends"]

[project.scripts]
ictye-live-dm = "ictye_live_dm.main:main"